requirements.txt
Django>=5.2,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
python-dotenv>=1.0
django-cors-headers>=4.6
Pillow>=10.0
